restructure web folder
