+++ website/src/plugins/pluginPlayground/web/CodeRunner/compiler/compileComponentFromFiles.ts
