+++ website/src/plugins/pluginPlayground/web/Playground/Playground.tsx
@@ -3,7 +3,7 @@ import clsx from 'clsx'
 import { useDark } from 'rspress/runtime'
 import { useFullscreen, useLocalStorage, useMediaQuery } from '@mantine/hooks'
 import { SandpackProvider, SandpackFiles } from '@codesandbox/sandpack-react'
-import { ENTRY_FILE_NAME } from '../../../shared/constants'
+import { ENTRY_FILE_NAME } from '../../shared/constants'
 import { Panels } from '../Panels/Panels'
 import { ControlPanel } from '../ControlPanel/ControlPanel'
 import { PanelsLayout } from '../constants'
