+++ website/src/plugins/pluginPlayground/web/Playground/index.ts
@@ -1,1 +1,1 @@
-export { Playground as default } from './Playground/Playground'
+export { Playground as default } from './Playground'
